@@ -1,17 +1,15 @@
 package com.example.controller;
 
-import com.example.ProjectsCache;
 import com.example.entity.Project;
 import com.example.entity.User;
 import com.example.service.ProjectService;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.annotation.WebServlet;
+import jakarta.servlet.http.HttpServlet;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.time.LocalDateTime;
 
 @WebServlet("/managerx/profile")
 public class ProfileServlet extends HttpServlet {
